+++ Runtime/AsyncCoroutineRunner.cs
@@ -14,8 +14,7 @@ namespace GameLovers.AsyncAwait
             {
                 if (_instance == null)
                 {
-                    _instance = new GameObject("AsyncCoroutineRunner")
-                        .AddComponent<AsyncCoroutineRunner>();
+                    _instance = new GameObject("AsyncCoroutineRunner").AddComponent<AsyncCoroutineRunner>();
                 }
 
                 return _instance;