+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2020-04-06
+
+- Added UnityWebRequestAwaiter. Now it is possible to do async await calls to UnityWebRequests
+- Added sample example to UnityWebRequestAwaiter
+
 ## [0.1.0] - 2020-01-06
 
 - Initial submission for package distribution